requirements.txt
# Core dependencies
pyspark==3.5.1
kafka-python==2.0.2
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.24.3

# Configuration and utilities
pyyaml==6.0.1
python-dotenv==1.0.0
joblib==1.3.2

# Real-time data sources
requests==2.31.0
websocket-client==1.6.4
yfinance==0.2.28

# Database dependencies
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Web dashboard
flask==3.0.0
flask-cors==4.0.0

# Logging and monitoring
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-mock==3.12.0

# Development
black==23.12.1
flake8==6.1.0 